chapter 1: WHAT THE FUCK and compiling ACC.
- setup VCPKG and cmake shit
- refactor a TON of headers misnamed
- renamed files to be like better and shit

uh yeah im using aya and projectpizza as references because holy shit this is a lot and takes a lot of time which sucks

i wanted to do studio originally but qt5 webkit is being a bitch and taking WAYY too long, my pc is ass but ill get to it

rn im targetting ACC since that has security and doesn't require as much libraries (cough Qt)

networking is almost done?? idk its a bunch of bullshit, the references are batshit insane and i want to keep security sooo yeah...

+++ Network/raknet/Source/RakNetTime.h
@@ -7,9 +7,9 @@
 namespace RakNet {
 
 // ROBLOX CHANGE: ALWAYS 64-bit
-typedef unsigned long long Time;
+typedef uint64_t Time;
 typedef unsigned int TimeMS;
-typedef unsigned long long TimeUS;
+typedef uint64_t TimeUS;
 // END ROBLOX CHANGE
 
 }; // namespace RakNet
